more shot end cleanups
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_99_shot_branch@17081 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/RemotePlayer.cxx
@@ -16,6 +16,7 @@
 /* common implementation headers */
 #include "World.h"
 #include "SyncClock.h"
+#include "playing.h"
 
 
 RemotePlayer::RemotePlayer(const PlayerId& _id, TeamColor _team,
@@ -25,7 +26,7 @@ RemotePlayer::RemotePlayer(const PlayerId& _id, TeamColor _team,
 {
   if (World::getWorld()) 
   {
-    for ( int i = 0; i < World::getWorld()->getMaxShots(); i++ )
+    for ( size_t i = 0; i < World::getWorld()->getMaxShots(); i++ )
       shotSlots.push_back(ShotSlot());
   }
  }
